requirements.txt
customtkinter
SpeechRecognition
pyaudio
google-genai